fixing some relative paths after the main folder move. #136.
