+++ package.json
@@ -2,7 +2,7 @@
   "name": "invest-workbench",
   "version": "0.1.0-alpha",
   "description": "Models that map and value the goods and services from nature that sustain and fulfill human life",
-  "main": "src/main.js",
+  "main": "src/main/main.js",
   "homepage": "./",
   "scripts": {
     "start": "electron -r @babel/register .",
